🐯完成保存SPU和SKU基本信息的功能

+++ gulimall-product/src/main/java/com/losgai/gulimall/product/vo/spus/Bounds.java
@@ -6,9 +6,11 @@ package com.losgai.gulimall.product.vo.spus;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Bounds {
 
-	private Integer buyBounds;
-	private Integer growBounds;
+	private BigDecimal buyBounds;
+	private BigDecimal growBounds;
 }
