+++ gulimall-product/src/main/java/com/losgai/gulimall/product/vo/spus/SpuSaveVo.java
@@ -6,15 +6,16 @@ package com.losgai.gulimall.product.vo.spus;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class SpuSaveVo {
 	private String spuName;
 	private String spuDescription;
-	private Integer catalogId;
-	private Integer brandId;
-	private Integer weight;
+	private Long catalogId;
+	private Long brandId;
+	private BigDecimal weight;
 	private Integer publishStatus;
 	private List<String> decript;
 	private List<String> images;
